+++ includes/class-tmsm-admin-cleanup.php
@@ -223,6 +223,9 @@ class Tmsm_Admin_Cleanup {
 		$this->loader->add_filter( 'wpo_wcpdf_template_styles', $plugin_admin, 'wpo_wcpdf_template_styles', 50, 2 );
 		$this->loader->add_filter( 'wpo_wcpdf_listing_actions', $plugin_admin, 'wpo_wcpdf_listing_actions', 50, 2 );
 
+		// ACF
+		$this->loader->add_action( 'acf/input/admin_footer', $plugin_admin, 'acf_input_disable_autocomplete' );
+
 	}
 
 	/**