+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 ### Unreleased ###
-
+* Move SmushIt to submenu
 
 ### 1.0.7: June 29nd, 2018
 * Temporary security fix