+++ CHANGELOG.md
@@ -1,8 +1,9 @@
-### Unreleased ###
+### 1.0.8: August 29th, 2018
 * Move SmushIt to submenu
 * ACF Disable Autocomplete
 * Change link to Customers admin menu
 * Remove admin footer
+* Elementor: display hint in shortcodes
 
 ### 1.0.7: June 29nd, 2018
 * Temporary security fix